test(container): adjust moved component and page tests
