--- apps/container/src/components/ConceptsPieChart/ConceptsPieChart.test.ts
@@ -1,22 +1,0 @@
-import { concepts } from '@/mocks/reviews';
-import { render, screen } from '@/test-utils';
-
-import { ConceptsPieChart } from '.';
-
-const total = Object.values(concepts).reduce((acc, curr) => acc + curr, 0);
-
-describe('<ConceptsPieChart />', () => {
-  test('render Concepts Pie Highchart', async () => {
-    render(ConceptsPieChart, {
-      props: {
-        grades: concepts,
-      },
-    });
-
-    Object.values(concepts).forEach((concept) => {
-      expect(
-        screen.getByText(RegExp(((100 * concept) / total).toFixed(1), 'i')),
-      ).toBeInTheDocument();
-    });
-  });
-});
