+++ apps/container/src/pages/Admin/AdminPage.test.ts
@@ -1,0 +1,10 @@
+import { render, screen } from '@/test-utils';
+
+import { AdminPage } from '.';
+
+describe('<AdminPage />', () => {
+  test('render admin view', () => {
+    render(AdminPage);
+    expect(screen.getByRole('heading', { name: 'Admin' })).toBeInTheDocument();
+  });
+});
