+++ apps/container/src/pages/Planning/PlanningPage.test.ts
@@ -1,0 +1,12 @@
+import { render, screen } from '@/test-utils';
+
+import { PlanningPage } from '.';
+
+describe('<PlanningPage />', () => {
+  test('render admin view', () => {
+    render(PlanningPage);
+    expect(
+      screen.getByRole('heading', { name: 'Planning' })
+    ).toBeInTheDocument();
+  });
+});
