--- apps/container/src/views/Admin/AdminView.test.ts
@@ -1,10 +1,0 @@
-import { render, screen } from '@/test-utils';
-
-import { AdminView } from '.';
-
-describe('<AdminView />', () => {
-  test('render admin view', () => {
-    render(AdminView);
-    expect(screen.getByRole('heading', { name: 'Admin' })).toBeInTheDocument();
-  });
-});
